[ZEPPELIN-546] Refactoring
