+++ zeppelin-interpreter/src/main/java/org/apache/zeppelin/dep/DependencyResolver.java
@@ -28,8 +28,6 @@ import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.sonatype.aether.RepositorySystem;
-import org.sonatype.aether.RepositorySystemSession;
 import org.sonatype.aether.artifact.Artifact;
 import org.sonatype.aether.collection.CollectRequest;
 import org.sonatype.aether.graph.Dependency;
@@ -45,48 +43,18 @@ import org.sonatype.aether.util.filter.PatternExclusionsDependencyFilter;
 
 /**
  * Deps resolver.
- * Add new dependencies from mvn repo (at runetime) to Zeppelin.
- *
+ * Add new dependencies from mvn repo (at runtime) to Zeppelin.
  */
-public class DependencyResolver {
+public class DependencyResolver extends AbstractDependencyResolver {
   Logger logger = LoggerFactory.getLogger(DependencyResolver.class);
-  private RepositorySystem system = Booter.newRepositorySystem();
-  private List<RemoteRepository> repos = new LinkedList<RemoteRepository>();
-  private RepositorySystemSession session;
 
   private final String[] exclusions = new String[] {"org.apache.zeppelin:zeppelin-zengine",
                                                     "org.apache.zeppelin:zeppelin-interpreter",
                                                     "org.apache.zeppelin:zeppelin-server"};
 
   public DependencyResolver(String localRepoPath) {
-    session = Booter.newRepositorySystemSession(system, localRepoPath);
-    repos.add(Booter.newCentralRepository()); // add maven central
-    repos.add(Booter.newLocalRepository());
+    super(localRepoPath);
   }
-
-  public void addRepo(String id, String url, boolean snapshot) {
-    synchronized (repos) {
-      delRepo(id);
-      RemoteRepository rr = new RemoteRepository(id, "default", url);
-      rr.setPolicy(snapshot, null);
-      repos.add(rr);
-    }
-  }
-
-  public RemoteRepository delRepo(String id) {
-    synchronized (repos) {
-      Iterator<RemoteRepository> it = repos.iterator();
-      if (it.hasNext()) {
-        RemoteRepository repo = it.next();
-        if (repo.getId().equals(id)) {
-          it.remove();
-          return repo;
-        }
-      }
-    }
-    return null;
-  }
-
 
   public List<File> load(String artifact) throws Exception {
     return load(artifact, new LinkedList<String>());
@@ -96,7 +64,8 @@ public class DependencyResolver {
     return load(artifact, new LinkedList<String>(), destPath);
   }
 
-  public List<File> load(String artifact, Collection<String> excludes) throws Exception {
+  public synchronized List<File> load(String artifact, Collection<String> excludes)
+      throws Exception {
     if (StringUtils.isBlank(artifact)) {
       // Should throw here
       throw new RuntimeException("Invalid artifact to load");
@@ -166,12 +135,12 @@ public class DependencyResolver {
   }
 
   /**
-   *
    * @param dependency
    * @param excludes list of pattern can either be of the form groupId:artifactId
    * @return
    * @throws Exception
    */
+  @Override
   public List<ArtifactResult> getArtifactsWithDep(String dependency,
       Collection<String> excludes) throws Exception {
     Artifact artifact = new DefaultArtifact(dependency);
